backend/SecondBrain.PythonApi/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.1
httpx>=0.25.1
pydantic-settings>=2.1.0

# AI and ML
openai>=1.3.3
anthropic>=0.5.0
google-generativeai>=0.3.0  # Google Generative AI package
langchain>=0.0.339
tiktoken>=0.5.1
nltk>=3.9.1
python-magic>=0.4.27

# Search and APIs
duckduckgo-search>=3.9.9
newsapi-python>=0.2.7
scholarly>=1.7.11
beautifulsoup4>=4.12.2

# Content Processing
PyMuPDF>=1.23.6  # fitz module for PDF processing
python-docx>=1.0.1  # for DOCX processing
pandas>=2.1.3  # for Excel/CSV processing
openpyxl>=3.1.2  # Excel support for pandas
lxml>=4.9.3  # XML/HTML processing
chardet>=5.2.0  # Character encoding detection

# Utilities
aiohttp>=3.9.1
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.2.1
tenacity>=8.2.3
cachetools>=5.3.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.1
